requirements.txt
chromadb
psycopg2-binary
numpy
psutil